requirements.txt
fastapi==0.115.13
joblib==1.5.1
matplotlib==3.10.3
numpy==2.0.2
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
pandas==2.3.0
pydantic==2.11.7
python-dotenv==1.1.0
rembg==2.0.66
scikit_learn==1.7.0
python-multipart==0.0.20
uvicorn==0.34.3
onnxruntime==1.22.0
gdown==4.7.1